Functional -> Function
Use the term "function" instead of "functional" when referring to React function components.

When hooks were introduced, SFC (Stateless Functional Component) was dropped in favor of FC (Function Component). With hooks, function components are no longer restricted to being stateless. The term "function" was adopted instead of "functional" since the latter implies Functional Programming paradigms such as pure functions, which is not a requirement of a function component.

https://twitter.com/dan_abramov/status/1057625147216220162
https://github.com/DefinitelyTyped/DefinitelyTyped/pull/30364